Split tests by model
