+++ bundles/org.dataflowanalysis.analysis.converter/src/org/dataflowanalysis/analysis/converter/microsecend/ExternalEntity.java
@@ -1,4 +1,4 @@
 package org.dataflowanalysis.analysis.converter.microsecend;
 
-public class ExternalEntity extends MicroSecEndProcess{
+public class ExternalEntity extends MicroSecEndProcess {
 }
