+++ bundles/org.dataflowanalysis.analysis.converter/src/org/dataflowanalysis/analysis/converter/microsecend/Service.java
@@ -1,4 +1,4 @@
 package org.dataflowanalysis.analysis.converter.microsecend;
 
-public class Service extends MicroSecEndProcess{
+public class Service extends MicroSecEndProcess {
 }
